SwingGroupingInputHarvester: Factor out SwingGroupingLabelListener
